All codes are now together and running
Not sure if the knn is working correctly (generates the same result
for all number of Xtest except the last two). This can be a problem
in the knn function or in the matrix creation functions, I am not sure.

The reading and writing are working correctly. If you have any issues
with the conversion, uncomment the "setlocale" line in the
readAndConvert.c file and recompile.

Use "make ep" to compile all the files needed for the full code to run
with no issues. If you need to test a single file, uncomment the main
and remove the tasks.json lines 10-12. Readd them to run or debug
fullCodeNoParallel.c.

Changes were made in all the .h files correcting the function calls,
the findArraySize function was deemed unnecessary and removed from
matrizes.c and matrizes.h.

Fixed issues on readAndConvert.c.

*.o files were added to .gitignore. .vscode was removed from gitignore
so the correct calls would be registered for debugging.

output/Ytest.txt was generated so the filepath for the Ytest file
works correctly. If you have issues with this, check if the filepath
in your computer matches the filepath in the readArrayAndConvertToFile
function in readAndConvert.c. If there are issues in this contact and
we'll add a user prompt for them to enter the correct directories (or
add the paths to the main args).

	modified:   .gitignore

	new file:   .vscode/settings.json

	new file:   .vscode/tasks.json

	new file:   Makefile

	modified:   fullCodeNoParallel.c

	modified:   knn.h

	modified:   matrizes.c

	modified:   matrizes.h

	modified:   output/Ytest.txt

	modified:   readAndConvert.c

	modified:   readAndConvert.h